examples/humaneval/pyproject.toml
[project]
name = "boxer-humaneval"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "boxer-sdk",
    "litellm>=1.0",
    "datasets>=2.0",
    "tqdm>=4.0",
    "python-dotenv>=1.0",
    "black>=26.3.1",
]

[dependency-groups]
dev = []

[tool.uv.sources]
boxer-sdk = { path = "../../packages/sdk/python" }
